Use reporting Activity in context menu bullet, rename ReportContext class
